[FEATURE] Validate task is running via Ajax
With this change it is possible to validate the status of a task via
Ajax. The default action is set to do so directly. With the call via
Ajax it is now possible to e.g. identify that route in an Apache rule
and redirect it to a specific server. This might be handy in a multi-
server environment where the cron tasks are only executed on one instance.